Improve extraction of certificate data from PEM files
Ignore content before and after the '---BEGIN <TYPE>---' and
'---END <TYPE>---' lines when extracting the base64 encoded
certificate from the PEM file for inclusion in the table of contents.

When following the instructions in the README to extract
certificates from PKCS 12 files, the exported PEM certificate
contains additional data before the 'BEGIN CERTIFICATE' line.