Fix addition of top-level files outside current directory
When generating an archive, `xarjs create` creates a XarFile tree from
the list of paths specified on the command line. Each of these entries
used contain only the name of the file, not its path.

eg. Given the path 'some-dir/root-dir' as an input file, walk() would
return:

  XarDirectory {
    name: 'root-dir'
    ...
    children: [
      XarFile {
        name: 'Info.plist',
      },
    ]
  }

When later reading the file source in order to generate the archive,
the concatenation of the names in the tree to FileReader, eg. 'root-dir'
and 'root-dir/Info.plist' in this example.

If one of the top-level input files was not in the current directory,
then the path would be invalid and reading it would fail. This went
unnoticed in tests becaused they used a `FileReader` instance which
resolved paths relative to the text fixture's root dir.

This commit fixes the issue by recording the absolute source path to the
file in XarFile entries.

Fixes #8